+++ src/main/java/com/shoppingorderapi/presentation/dto/auth/request/SignUpRequestDto.java
@@ -1,16 +1,22 @@
-package com.shoppingorderapi.presentation.dto.user.request;
+package com.shoppingorderapi.presentation.dto.auth.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
 
 import lombok.Getter;
 
 @Getter
-public class SignInRequestDto {
+public class SignUpRequestDto {
+
 	@NotBlank(message = "이메일은 필수입니다")
 	@Email(message = "올바른 이메일 형식이 아닙니다")
 	private String email;
 
 	@NotBlank(message = "비밀번호는 필수입니다")
+	@Size(min = 8, message = "비밀번호는 최소 8자 이상이어야 합니다")
 	private String password;
+
+	@NotBlank(message = "주소는 필수입니다")
+	private String address;
 }
